@@ -2,13 +2,15 @@ const express = require("express");
 const verifyToken = require("./utils/verifyToken");
 
 const BookController = require("./controllers/BookController");
-const UserController = require("./controllers/UserController");
-const SessionController = require("./controllers/SessionController");
-const ChapterController = require("./controllers/ChapterController");
+const VerseController = require("./controllers/VerseController");
 const CommentController = require("./controllers/CommentController");
-const SearchController = require("./controllers/SearchController");
-const DiscussionController = require("./controllers/DiscussionController");
+
+const UserController = require("./controllers/UserController");
 const InfoController = require("./controllers/InfoController");
+const SessionController = require("./controllers/SessionController");
+
+const DiscussionController = require("./controllers/DiscussionController");
+const SearchController = require("./controllers/SearchController");
 const BackupController = require("./controllers/BackupController");
 
 const routes = express.Router();
@@ -39,21 +41,14 @@ routes.get("/books", BookController.index);
 routes.post("/books", BookController.store);
 routes.get("/books/:abbrev", BookController.show);
 
-routes.get("/books/:abbrev/chapters", ChapterController.index);
-routes.get("/books/:abbrev/chapters/:number", ChapterController.show);
-routes.post("/books/:abbrev/chapters/:number", ChapterController.store);
-routes.patch("/books/:abbrev/chapters/:number", ChapterController.update);
+routes.get("/books/:abbrev/verses", VerseController.index);
+routes.get("/books/:abbrev/verses/:chapter", VerseController.show);
+routes.post("/books/:abbrev/verses/:chapter/:verse", VerseController.store);
+routes.patch("/books/:abbrev/verses/:chapter/:verse", VerseController.update);
 
-routes.get("/books/:abbrev/chapters/:number/comments", CommentController.index);
-routes.get(
-	"/books/:abbrev/chapters/:number/comments/:verse",
-	CommentController.show
-);
-routes.post(
-	"/books/:abbrev/chapters/:number/comments/:verse",
-	Authentication,
-	CommentController.store
-);
+routes.get("/comments/:abbrev/:chapter", CommentController.index);
+routes.get("/comments/:abbrev/:chapter/:verse", CommentController.show);
+routes.post("/comments/:verseID", Authentication, CommentController.store);
 routes.patch("/comments/:id", Authentication, CommentController.update);
 routes.delete("/comments/:id", Authentication, CommentController.destroy);
 
