@@ -1,14 +1,11 @@
 import PropTypes from "prop-types";
 
 export default PropTypes.shape({
-	tags: PropTypes.oneOfType([
-		PropTypes.string,
-		PropTypes.arrayOf(PropTypes.string),
-	]).isRequired,
-	verse: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
 	id: PropTypes.number.isRequired,
 	text: PropTypes.string.isRequired,
-	likes: PropTypes.string.isRequired,
+	verse_id: PropTypes.number.isRequired,
+	tags: PropTypes.arrayOf(PropTypes.string).isRequired,
+	likes: PropTypes.arrayOf(PropTypes.string).isRequired,
 	on_title: PropTypes.oneOfType([PropTypes.number, PropTypes.bool]).isRequired,
 	username: PropTypes.string.isRequired,
 	created_at: PropTypes.string.isRequired,
